Hữu Tài - Xong 70% CategoriesController còn lỗi update khi Categories nằm trong books

+++ BookWorld/src/main/java/com/poly/dao/AuthorBookDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.AuthorBook;
 
-public interface AuthorBookDAO extends JpaRepository<AuthorBook,Integer> {
+public interface AuthorBookDAO extends JpaRepository<AuthorBook,Long> {
     
 }
