+++ BookWorld/src/main/java/com/poly/dao/AuthorDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.Author;
 
-public interface AuthorDAO extends JpaRepository <Author,Integer> {
+public interface AuthorDAO extends JpaRepository <Author,Long> {
     
 }
