+++ BookWorld/src/main/java/com/poly/dao/CategoryDAO.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.Category;
 
-public interface CategoryDAO extends JpaRepository<Category,Integer> {
+public interface CategoryDAO extends JpaRepository<Category,Long> {
+
     
 }
