+++ BookWorld/src/main/java/com/poly/dao/OrderDetailDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.OrderDetail;
 
-public interface OrderDetailDAO extends JpaRepository <OrderDetail,Integer> {
+public interface OrderDetailDAO extends JpaRepository <OrderDetail,Long> {
     
 }
