+++ BookWorld/src/main/java/com/poly/dao/PublisherDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.Publisher;
 
-public interface PublisherDAO extends JpaRepository<Publisher,Integer> {
+public interface PublisherDAO extends JpaRepository<Publisher,Long> {
     
 }
