+++ BookWorld/src/main/java/com/poly/dao/ShareDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poly.model.Share;
 
-public interface ShareDAO extends JpaRepository <Share,Integer> {
+public interface ShareDAO extends JpaRepository <Share,Long> {
     
 }
